Move the AI Image Editor front end out of toolbox/canvas (BL-16682)
Pure relocation plus import paths; no behavior change. Review with `git show -M`
and it is 11 renames, four of which have no content change at all.

WHY

`bookEdit/toolbox/canvas/AGENTS.md` opens with "This folder contains Canvas Tool
UI and related Canvas utilities", and its README describes "the declarative system
that controls what operations are offered for different kinds of canvas
elements". The AI image editor is neither, and it had grown to 11 of the 51 files
in there -- the largest non-canvas cluster in the folder.

It is also not a toolbox tool: every other `toolbox/*` folder is an actual tool
with a panel and a lifecycle, whereas this is a context-menu command plus a
dialog-like overlay. That makes `bookEdit/copyrightAndLicense/` its closest
structural neighbour, so it now sits beside it as `bookEdit/aiImageEditor/`. The
parallel is exact rather than loose: both are image-related, both are launched
from the same canvas context menu, and both render in the top window precisely so
a page save cannot tear them down.

The coupling to canvas is two imports, both from the menu machinery, which is the
right relationship -- the canvas menu offers the feature, it does not implement
it: canvasControlRegistry imports launchAiImageEditor, and
buildCanvasElementControlRegistryContext imports isAiEditableImageSrc to decide
whether the item is enabled.

WHAT THE MOVE BUYS BEYOND TIDINESS

A folder of its own is somewhere to write down the frame rule, which until now
lived only in individual file headers: the overlay half must run in the top
window and the live-page half in the page iframe, they must never import each
other, and anything shared must stay pure. Bundling follows imports rather than
folders, so nothing else enforces that. The new AGENTS.md states it, with the
table of which file belongs to which frame, and points at HandleSaveThenLaunch
for why waiting for the page load is also required. AiImageEditorApi.cs now names
the folder so the C# side leads a reader to it.

WHAT NEEDED NO CHANGE

- Playwright: `playwright.bloom-exe.config.ts` uses testDir ".." with testMatch
  "**/bloom-exe*.uitest.ts" -- a glob, and its own comment says the setup exists
  "so a feature's e2e can live next to its code".
- Vitest: its include is likewise a path-agnostic glob.
- Every cross-reference in the C# and the CDP driving guide cites file *names*,
  not paths, and the one that mentions a path already elides it.

VERIFICATION

- typecheck, lint clean. The moved suites are found at the new path and pass:
  aiEditorOverlay (10), aiEditorPageCommands (6), aiEditorSlotMatching (5),
  aiEditorImageFormats (4) -- 35 with the canvas suites, same as before.
- The isolated production bundle builds (build/agent-vite.sh).
- BloomExe builds clean.

A note on what was NOT verified, so nobody assumes otherwise: I tried to confirm
at the bundle level that the two halves still land in the right places, and that
check is inconclusive here -- this build funnels shared code into one 4.5MB chunk
imported by eight entry points, so a grep cannot distinguish them. It is also
unnecessary: the move changed no import edges, only paths, so bundle composition
cannot have shifted. The guarantee that matters is the source import graph, which
is still one-way, and which the new AGENTS.md now documents.
