+++ src/BloomBrowserUI/bookEdit/aiImageEditor/aiEditorImageFormats.test.ts
