+++ src/BloomBrowserUI/bookEdit/aiImageEditor/aiEditorImageFormats.ts
