+++ src/BloomBrowserUI/bookEdit/aiImageEditor/aiEditorShared.ts
